PDFBOX-3065: remove deprecated constructor; remove unused variables; remove super(); integrate initWriter() into constructor; change COSStandardOutputStream constructor parameter
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1760167 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/pdfwriter/COSStandardOutputStream.java
@@ -64,7 +64,7 @@ public class COSStandardOutputStream extends FilterOutputStream
      * @param out The underlying stream to write to.
      * @param position The current position of output stream.
      */
-    public COSStandardOutputStream(OutputStream out, int position)
+    public COSStandardOutputStream(OutputStream out, long position)
     {
         super(out);
         this.position = position;
